+++ tests/ObjectIndexerTest.php
@@ -105,7 +105,7 @@ class ObjectIndexerTest extends TestCase
     {
         $this->expectException(UndefinedKeyException::class);
 
-        $this->expectErrorMessage('Undefined index key 1');
+        $this->expectExceptionMessage('Undefined index key 1');
 
         $dico = new PropertyIndexer('id', 'value');
         $dico->get(1);